CAL-412 Updated Cloud Cover to be a float (#529)
* CAL-412 Made Cloud Cover a float type

* CAL-412 Code Review and some clean up

+++ catalog/imaging/imaging-transformer-nitf/src/main/java/org/codice/alliance/transformer/nitf/common/PiaimcAttribute.java
@@ -288,7 +288,7 @@ public class PiaimcAttribute extends NitfAttributeImpl<Tre> {
         .filter(String.class::isInstance)
         .map(String.class::cast)
         .filter(StringUtils::isNotEmpty)
-        .map(Integer::valueOf)
+        .map(Double::valueOf)
         .filter(value -> value >= 0 && value <= 100)
         .orElse(null);
   }
